backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
python-multipart
openai==2.0.1